requirements.txt
fastapi
uvicorn
neo4j
pinecone-client
requests
python-dotenv
pydantic
tqdm
